Remake saving
The system in the previous commit wasn't right.
More than 50% of the data wasn't used.
That brought additional bugs with it.

These are now solved. Saving and loading is working with the new system (based on the old) and uses the available data more efficient.
The new boards need to be made and saved.
Further testing is necessary.

Extra Key command for making, cleaning, saving,... the board are added.